{"schema":"decodex/commit/1","summary":"Modularize orchestrator run loop","authority":"manual"}
