--- apps/decodex/src/orchestrator/daemon.rs
@@ -1,950 +1,0 @@
-use crate::cli::AttemptRequest;
-use daemon_retry::{
-	clear_retry_schedule_and_release, retry_entry_is_temporarily_blocked,
-	schedule_retry_after_child_exit,
-};
-
-struct DaemonTickRuntimeContext<'a, T, I> {
-	tracker: &'a T,
-	project: &'a ServiceConfig,
-	workflow: &'a WorkflowDocument,
-	worktree_manager: &'a WorktreeManager,
-	review_state_inspector: &'a I,
-	recoverable_worktree_skip_cache: Option<&'a mut RecoverableWorktreeSkipCache>,
-}
-
-fn load_daemon_tick_context(
-	config_path: &Path,
-	workflow_cache: &mut Option<CachedWorkflowDocument>,
-) -> Result<DaemonTickContext> {
-	let config = ServiceConfig::from_path(config_path)?;
-	let workflow = load_daemon_tick_workflow(&config, workflow_cache)?;
-	let api_key = config.tracker().resolve_api_key()?;
-	let tracker = LinearClient::new(api_key)?;
-	let worktree_manager =
-		WorktreeManager::new(config.service_id(), config.repo_root(), config.worktree_root());
-
-	Ok(DaemonTickContext { config, workflow, tracker, worktree_manager })
-}
-
-fn load_daemon_tick_workflow(
-	config: &ServiceConfig,
-	workflow_cache: &mut Option<CachedWorkflowDocument>,
-) -> Result<WorkflowDocument> {
-	let workflow_path = config.workflow_path().to_path_buf();
-	let cached_same_path = workflow_cache
-		.as_ref()
-		.filter(|cached| cached.path == workflow_path)
-		.map(|cached| cached.document.clone());
-
-	match WorkflowDocument::from_path(&workflow_path) {
-		Ok(workflow) => {
-			if cached_same_path.as_ref().is_some_and(|cached| cached != &workflow) {
-				tracing::info!(
-					workflow_path = %workflow_path.display(),
-					"Reloaded project WORKFLOW.md for future control-plane decisions."
-				);
-			}
-
-			*workflow_cache =
-				Some(CachedWorkflowDocument { path: workflow_path, document: workflow.clone() });
-
-			Ok(workflow)
-		},
-		Err(error) => {
-			if let Some(cached_workflow) = cached_same_path {
-				tracing::warn!(
-					workflow_path = %workflow_path.display(),
-					?error,
-					"Failed to reload project WORKFLOW.md; keeping the last known good workflow active for control-plane decisions."
-				);
-
-				Ok(cached_workflow)
-			} else {
-				Err(error)
-			}
-		},
-	}
-}
-
-fn run_daemon_tick(
-	config_path: &Path,
-	state_store: &StateStore,
-	active_children: &mut Vec<DaemonRunChild>,
-	retry_queue: &mut RetryQueue,
-	recoverable_worktree_skip_cache: &mut RecoverableWorktreeSkipCache,
-	context: &DaemonTickContext,
-) -> Result<()> {
-	let review_state_inspector = GhPullRequestReviewStateInspector {
-		github_token_env_var: Some(context.config.github().token_env_var().to_owned()),
-		github_command_path: context.config.github().command_path().map(Path::to_path_buf),
-	};
-
-	run_daemon_tick_with_review_state_inspector(
-		config_path,
-		state_store,
-		active_children,
-		retry_queue,
-		DaemonTickRuntimeContext {
-			tracker: &context.tracker,
-			project: &context.config,
-			workflow: &context.workflow,
-			worktree_manager: &context.worktree_manager,
-			review_state_inspector: &review_state_inspector,
-			recoverable_worktree_skip_cache: Some(recoverable_worktree_skip_cache),
-		},
-	)
-}
-
-fn run_daemon_tick_with_review_state_inspector<T, I>(
-	config_path: &Path,
-	state_store: &StateStore,
-	active_children: &mut Vec<DaemonRunChild>,
-	retry_queue: &mut RetryQueue,
-	mut context: DaemonTickRuntimeContext<'_, T, I>,
-) -> Result<()>
-where
-	T: IssueTracker,
-	I: PullRequestReviewStateInspector,
-{
-	inspect_or_clear_active_children(
-		active_children,
-		retry_queue,
-		context.tracker,
-		context.project,
-		context.workflow,
-		state_store,
-		context.worktree_manager,
-	)?;
-
-	if active_children.is_empty() {
-		let recoverable_worktree_skip_cache =
-			context.recoverable_worktree_skip_cache.as_deref_mut();
-
-		recover_and_reconcile_idle_daemon_state(
-			context.tracker,
-			context.project,
-			context.workflow,
-			state_store,
-			context.worktree_manager,
-			recoverable_worktree_skip_cache,
-		)?;
-	}
-
-	reconcile_post_review_orchestration_with_inspector(
-		context.tracker,
-		context.project,
-		context.workflow,
-		state_store,
-		context.review_state_inspector,
-	)?;
-	reconcile_terminal_thread_archive_backlog_best_effort(
-		context.project,
-		context.workflow,
-		state_store,
-	);
-
-	loop {
-		if !spawn_next_daemon_child(
-			config_path,
-			state_store,
-			active_children,
-			retry_queue,
-			&context,
-		)? {
-			break;
-		}
-	}
-
-	Ok(())
-}
-
-fn recover_and_reconcile_idle_daemon_state<T>(
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	worktree_manager: &WorktreeManager,
-	recoverable_worktree_skip_cache: Option<&mut RecoverableWorktreeSkipCache>,
-) -> Result<()>
-where
-	T: IssueTracker,
-{
-	let _ = recover_runtime_state_from_tracker_and_worktrees_with_skip_cache(
-		tracker,
-		project,
-		workflow,
-		state_store,
-		recoverable_worktree_skip_cache,
-	)?;
-
-	reconcile_project_state(tracker, project, workflow, state_store, worktree_manager)
-}
-
-fn build_operator_state_snapshot_for_publish<T>(
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	limit: usize,
-	warnings: &[&str],
-	connector_backoffs: &[OperatorConnectorBackoffStatus],
-) -> Result<OperatorStatusSnapshot>
-where
-	T: IssueTracker,
-{
-	let mut snapshot = if warnings.is_empty() {
-		build_control_plane_operator_status_snapshot(
-			tracker,
-			project,
-			workflow,
-			state_store,
-			limit,
-		)?
-	} else {
-		build_operator_status_snapshot_with_account_mode(
-			project,
-			state_store,
-			limit,
-			AccountActivityMode::Snapshot,
-		)?
-	};
-
-	if !warnings.is_empty() {
-		hydrate_history_lanes_from_local_ledger(project, state_store, &mut snapshot)?;
-	}
-
-	apply_terminal_history_ledger_outcomes(&mut snapshot);
-
-	if warnings_include_tracker_backoff(warnings) {
-		let review_state_inspector = GhPullRequestReviewStateInspector {
-			github_token_env_var: Some(project.github().token_env_var().to_owned()),
-			github_command_path: project.github().command_path().map(Path::to_path_buf),
-		};
-
-		snapshot.post_review_lanes = build_degraded_post_review_lane_statuses(
-			project,
-			state_store,
-			&review_state_inspector,
-		)?;
-	}
-
-	for warning in warnings {
-		add_operator_snapshot_warning(&mut snapshot, warning);
-	}
-
-	snapshot.connector_backoffs.extend(connector_backoffs.iter().cloned());
-
-	if !warnings.is_empty() {
-		add_operator_snapshot_warning(&mut snapshot, "external_observer_status_skipped");
-	}
-
-	refresh_operator_project_summary(&mut snapshot, None);
-
-	Ok(snapshot)
-}
-
-fn inspect_or_clear_active_children<T>(
-	active_children: &mut Vec<DaemonRunChild>,
-	retry_queue: &mut RetryQueue,
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	worktree_manager: &WorktreeManager,
-) -> Result<()>
-where
-	T: IssueTracker,
-{
-	let mut index = 0;
-
-	while index < active_children.len() {
-		let child_exit_status = active_children[index].child.try_wait()?;
-		let child_exited = child_exit_status.is_some();
-
-		if child_exited && child_exit_status.is_some_and(|status| !status.success()) {
-			mark_run_attempt_if_active(state_store, &active_children[index].run_id, "failed")?;
-		}
-
-		let child_ref = ChildRunRef {
-			issue_id: &active_children[index].issue_id,
-			run_id: &active_children[index].run_id,
-			attempt_number: active_children[index].attempt_number,
-		};
-		let actions = if child_exited {
-			inspect_exited_daemon_child_reconciliation(
-				tracker,
-				project,
-				workflow,
-				state_store,
-				child_ref.issue_id,
-				child_ref.run_id,
-			)?
-		} else {
-			inspect_current_daemon_child_reconciliation(
-				tracker,
-				project,
-				workflow,
-				state_store,
-				CurrentChildRunContext {
-					child: child_ref,
-					workflow: &active_children[index].workflow,
-					dispatch_mode: active_children[index].dispatch_mode,
-				},
-			)?
-		};
-
-		if actions.is_empty() {
-			if child_exited {
-				if child_exit_status.is_some_and(|status| status.success()) {
-					mark_run_attempt_if_active(
-						state_store,
-						&active_children[index].run_id,
-						"succeeded",
-					)?;
-				}
-
-				let daemon_child = active_children.swap_remove(index);
-				let child_ref = ChildRunRef {
-					issue_id: &daemon_child.issue_id,
-					run_id: &daemon_child.run_id,
-					attempt_number: daemon_child.attempt_number,
-				};
-
-				clear_orphaned_daemon_child_state(state_store, child_ref, false)?;
-
-				if let Some(exit_status) = child_exit_status {
-					schedule_retry_after_child_exit(
-						ChildExitRetryContext {
-							retry_queue,
-							tracker,
-							project,
-							workflow,
-							state_store,
-						},
-						child_ref,
-						#[cfg(test)]
-						"",
-						&daemon_child.initial_issue_state,
-						daemon_child.dispatch_mode,
-						exit_status,
-					)?;
-				}
-
-				continue;
-			}
-
-			index += 1;
-
-			continue;
-		}
-
-		let mut daemon_child = active_children.swap_remove(index);
-
-		if daemon_child.from_retry_queue {
-			retry_queue.release(&daemon_child.issue_id);
-		}
-		if !child_exited {
-			stop_daemon_child(&mut daemon_child.child)?;
-		}
-
-		apply_run_lease_reconciliation(tracker, project, state_store, worktree_manager, actions)?;
-	}
-
-	Ok(())
-}
-
-fn inspect_current_daemon_child_reconciliation<T>(
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	child_context: CurrentChildRunContext<'_>,
-) -> Result<Vec<RunLeaseReconciliation>>
-where
-	T: IssueTracker,
-{
-	inspect_current_daemon_child_reconciliation_at(
-		tracker,
-		project,
-		workflow,
-		state_store,
-		child_context,
-		OffsetDateTime::now_utc().unix_timestamp(),
-	)
-}
-
-fn inspect_current_daemon_child_reconciliation_at<T>(
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	child_context: CurrentChildRunContext<'_>,
-	now_unix_epoch: i64,
-) -> Result<Vec<RunLeaseReconciliation>>
-where
-	T: IssueTracker,
-{
-	let child = child_context.child;
-	let Some(issue) = refresh_issue(tracker, child.issue_id)? else {
-		return Ok(Vec::new());
-	};
-	let Some(run_attempt) = state_store.run_attempt(child.run_id)? else {
-		return Ok(Vec::new());
-	};
-	let worktree_mapping = state_store.worktree_for_issue(&issue.id)?;
-
-	if let Some(disposition) = superseded_run_disposition(state_store, &run_attempt)? {
-		return Ok(vec![RunLeaseReconciliation {
-			issue: issue.clone(),
-			run_attempt,
-			worktree_mapping,
-			disposition,
-			workflow: workflow.clone(),
-		}]);
-	}
-
-	let action_workflow = run_lease_reconciliation_workflow(
-		workflow,
-		Some(ActiveWorkflowOverride { child, workflow: child_context.workflow }),
-		&issue,
-		&run_attempt,
-	);
-	let retained_closeout = terminal_issue_keeps_retained_closeout(
-		tracker,
-		&issue,
-		project,
-		action_workflow,
-		state_store,
-	)?;
-	let completed_closeout_child =
-		matches!(child_context.dispatch_mode, IssueDispatchMode::Closeout)
-			&& is_terminal_issue(&issue, action_workflow);
-	let disposition = if !retained_closeout
-		&& !completed_closeout_child
-		&& is_terminal_issue(&issue, action_workflow)
-	{
-		Some(RunLeaseDisposition::Terminal)
-	} else if !retained_closeout
-		&& !completed_closeout_child
-		&& is_issue_not_dispatchable_for_current_dispatch(
-			tracker,
-			&issue,
-			project,
-			action_workflow,
-			child_context.dispatch_mode,
-		)? {
-		Some(RunLeaseDisposition::NotDispatchable)
-	} else if let Some(idle_for) =
-		stalled_idle_duration(state_store, &run_attempt, worktree_mapping.as_ref(), now_unix_epoch)?
-	{
-		if retained_review_handoff_matches_run(
-			state_store,
-			&run_attempt,
-			worktree_mapping.as_ref(),
-		)? {
-			Some(RunLeaseDisposition::RetainedReviewComplete)
-		} else if stalled_run_has_retained_partial_progress(worktree_mapping.as_ref()) {
-			Some(RunLeaseDisposition::StalledRetainedPartialProgress { idle_for })
-		} else {
-			Some(RunLeaseDisposition::Stalled { idle_for })
-		}
-	} else {
-		None
-	};
-
-	Ok(disposition.map_or_else(Vec::new, |disposition| {
-		vec![RunLeaseReconciliation {
-			issue: issue.clone(),
-			run_attempt,
-			worktree_mapping,
-			disposition,
-			workflow: action_workflow.clone(),
-		}]
-	}))
-}
-
-fn clear_orphaned_daemon_child_state(
-	state_store: &StateStore,
-	child: ChildRunRef<'_>,
-	mark_interrupted: bool,
-) -> Result<()> {
-	let resolved_run_attempt = resolve_child_exit_run_attempt(state_store, child)?;
-
-	if resolved_run_attempt.is_none() {
-		tracing::debug!(
-			issue_id = child.issue_id,
-			run_id = child.run_id,
-			attempt = child.attempt_number,
-			"Daemon child exited without a matching recorded run attempt; skipping orphan cleanup."
-		);
-	}
-	if mark_interrupted && let Some(run_attempt) = resolved_run_attempt.as_ref() {
-		mark_run_attempt_if_active(state_store, run_attempt.run_id(), "interrupted")?;
-	}
-
-	let existing_lease = state_store.lease_for_issue(child.issue_id)?;
-	let issue_unowned_or_matches_run = existing_lease.as_ref().is_none_or(|lease| {
-		resolved_run_attempt
-			.as_ref()
-			.is_some_and(|run_attempt| lease.run_id() == run_attempt.run_id())
-			|| lease.run_id() == child.run_id
-	});
-
-	if existing_lease.is_some() && issue_unowned_or_matches_run {
-		state_store.clear_lease(child.issue_id)?;
-	}
-	if resolved_run_attempt.is_some()
-		&& issue_unowned_or_matches_run
-		&& let Some(mapping) = state_store.worktree_for_issue(child.issue_id)?
-		&& !mapping.worktree_path().try_exists()?
-	{
-		tracing::debug!(
-			issue_id = child.issue_id,
-			run_id = child.run_id,
-			attempt = child.attempt_number,
-			branch = mapping.branch_name(),
-			worktree_path = %mapping.worktree_path().display(),
-			"Cleared daemon child worktree mapping after the checkout was removed."
-		);
-
-		state_store.clear_worktree(child.issue_id)?;
-	}
-
-	Ok(())
-}
-
-pub(super) fn resolve_child_exit_run_attempt(
-	state_store: &StateStore,
-	child: ChildRunRef<'_>,
-) -> Result<Option<RunAttempt>> {
-	state_store.run_attempt(child.run_id)
-}
-
-fn spawn_next_daemon_child<T>(
-	config_path: &Path,
-	state_store: &StateStore,
-	active_children: &mut Vec<DaemonRunChild>,
-	retry_queue: &mut RetryQueue,
-	context: &DaemonTickRuntimeContext<'_, T, impl PullRequestReviewStateInspector>,
-) -> Result<bool>
-where
-	T: IssueTracker,
-{
-	let next_run = plan_next_daemon_run(
-		retry_queue,
-		context.tracker,
-		context.project,
-		context.workflow,
-		state_store,
-	)?;
-
-	match next_run {
-		Some((summary, from_retry_queue)) => {
-			if summary.dispatch_mode != IssueDispatchMode::Closeout {
-				ensure_project_has_no_merged_worktree_cleanup_debt(context.project)?;
-			}
-
-			validate_workflow_read_first_files(context.project, context.workflow)?;
-
-			state_store.configure_dispatch_slot_root(
-				context.project.service_id(),
-				context.project.worktree_root(),
-			)?;
-
-			if !state_store.try_acquire_lease(
-				context.project.service_id(),
-				&summary.issue_id,
-				&summary.run_id,
-				&summary.issue_state,
-			)? {
-				return Ok(false);
-			}
-
-			let daemon_spawn_state = materialize_daemon_spawn_state(
-				context.project,
-				context.workflow,
-				state_store,
-				&summary,
-			)
-			.inspect_err(|_error| {
-				let _ = state_store.clear_lease(&summary.issue_id);
-			})?;
-
-			state_store.record_run_attempt(
-				&summary.run_id,
-				&summary.issue_id,
-				summary.attempt_number,
-				"starting",
-			)?;
-			state_store.upsert_worktree(
-				context.project.service_id(),
-				&summary.issue_id,
-				&daemon_spawn_state.worktree.branch_name,
-				&daemon_spawn_state.worktree.path.display().to_string(),
-			)?;
-
-			let mut child = spawn_planned_daemon_child(
-				config_path,
-				state_store,
-				context.workflow,
-				&summary,
-				daemon_spawn_state.retry_budget_base,
-			)?;
-
-			if let Err(error) = state::write_run_operation_marker_for_process(
-				&daemon_spawn_state.worktree.path,
-				&summary.run_id,
-				summary.attempt_number,
-				child.id(),
-				RUN_OPERATION_AGENT_RUN,
-			) {
-				let _ = child.kill();
-				let _ = child.wait();
-				let _ = state_store.update_run_status(&summary.run_id, "failed");
-				let _ = state_store.clear_lease(&summary.issue_id);
-
-				return Err(error);
-			}
-
-			state_store.update_run_status(&summary.run_id, "running")?;
-
-			tracing::info!(
-				issue = summary.issue_identifier,
-				worktree = %daemon_spawn_state.worktree.path.display(),
-				retry = from_retry_queue,
-				"Spawned control-plane child for current issue lane."
-			);
-
-			active_children.push(DaemonRunChild {
-				child,
-				issue_id: summary.issue_id,
-				run_id: summary.run_id,
-				attempt_number: summary.attempt_number,
-				initial_issue_state: summary.initial_issue_state,
-				#[cfg(test)]
-				retry_project_slug: String::new(),
-				dispatch_mode: summary.dispatch_mode,
-				from_retry_queue,
-				workflow: context.workflow.clone(),
-			});
-
-			Ok(true)
-		},
-		None => {
-			if retry_queue.is_empty() {
-				tracing::debug!("Daemon tick found no eligible issue.");
-			} else {
-				tracing::debug!("Daemon tick is holding a queued retry claim.");
-			}
-
-			Ok(false)
-		},
-	}
-}
-
-fn spawn_planned_daemon_child(
-	config_path: &Path,
-	state_store: &StateStore,
-	workflow: &WorkflowDocument,
-	summary: &RunSummary,
-	retry_budget_base: i64,
-) -> Result<Child> {
-	let issue_claim_handoff =
-		Some(state_store.clone_issue_claim_for_child(&summary.issue_id).inspect_err(|_error| {
-			let _ = state_store.update_run_status(&summary.run_id, "failed");
-			let _ = state_store.clear_lease(&summary.issue_id);
-		})?);
-	let (dispatch_slot_handoff_file, dispatch_slot_index) =
-		state_store.clone_dispatch_slot_for_child(&summary.issue_id)?;
-	let dispatch_slot_handoff = Some(dispatch_slot_handoff_file);
-	let dispatch_slot_index_handoff = Some(dispatch_slot_index);
-	let mut child = spawn_run_once_child(SpawnRunOnceChildRequest {
-		config_path,
-		preferred_issue_id: summary.issue_id.as_str(),
-		preferred_issue_state: summary.issue_state.as_str(),
-		preferred_initial_issue_state: Some(summary.initial_issue_state.as_str()),
-		dispatch_mode: summary.dispatch_mode,
-		preferred_run_id: summary.run_id.as_str(),
-		preferred_attempt_number: summary.attempt_number,
-		preferred_retry_budget_base: retry_budget_base,
-		workflow,
-		issue_claim_handoff: issue_claim_handoff.as_ref(),
-		dispatch_slot_handoff: dispatch_slot_handoff.as_ref(),
-		dispatch_slot_index_handoff,
-	})
-	.inspect_err(|_error| {
-		let _ = state_store.update_run_status(&summary.run_id, "failed");
-		let _ = state_store.clear_lease(&summary.issue_id);
-	})?;
-
-	state_store.release_handed_off_guards(&summary.issue_id).inspect_err(|_error| {
-		let _ = child.kill();
-		let _ = child.wait();
-		let _ = state_store.update_run_status(&summary.run_id, "failed");
-		let _ = state_store.clear_lease(&summary.issue_id);
-	})?;
-
-	Ok(child)
-}
-
-fn plan_next_daemon_run<T>(
-	retry_queue: &mut RetryQueue,
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-) -> Result<Option<(RunSummary, bool)>>
-where
-	T: IssueTracker,
-{
-	match plan_due_retry_run(retry_queue, tracker, project, workflow, state_store)? {
-		RetryDispatchDecision::Dispatch(summary) => Ok(Some((*summary, true))),
-		RetryDispatchDecision::Blocked { excluded_issue_ids } => {
-			let excluded_issue_ids =
-				excluded_issue_ids.iter().map(String::as_str).collect::<Vec<_>>();
-			let issue_run = plan_project_issue_run_with_exclusions(
-				tracker,
-				project,
-				workflow,
-				state_store,
-				true,
-				&excluded_issue_ids,
-			)?;
-
-			Ok(issue_run.map(|issue_run| {
-				(run_summary_from_issue_run(project.service_id(), &issue_run), false)
-			}))
-		},
-		RetryDispatchDecision::Continue => {
-			let issue_run = plan_project_issue_run_with_exclusions(
-				tracker,
-				project,
-				workflow,
-				state_store,
-				true,
-				&[],
-			)?;
-
-			Ok(issue_run.map(|issue_run| {
-				(run_summary_from_issue_run(project.service_id(), &issue_run), false)
-			}))
-		},
-	}
-}
-
-fn materialize_daemon_spawn_state(
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	summary: &RunSummary,
-) -> Result<MaterializedDaemonSpawnState> {
-	let worktree = materialize_run_summary_worktree(project, workflow, summary)?;
-	let retry_budget_base = retry_budget_base_for_dispatch_mode(
-		state_store,
-		&summary.issue_id,
-		&worktree.path,
-		summary.dispatch_mode,
-		None,
-	)?;
-
-	Ok(MaterializedDaemonSpawnState { worktree, retry_budget_base })
-}
-
-fn materialize_run_summary_worktree(
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	summary: &RunSummary,
-) -> Result<WorktreeSpec> {
-	if summary.dispatch_mode == IssueDispatchMode::Closeout {
-		if !summary.worktree_path.try_exists()? {
-			eyre::bail!(
-				"planned retained closeout worktree `{}` is missing for issue `{}`",
-				summary.worktree_path.display(),
-				summary.issue_identifier
-			);
-		}
-
-		return Ok(WorktreeSpec {
-			branch_name: summary.branch_name.clone(),
-			issue_identifier: summary.issue_identifier.clone(),
-			path: summary.worktree_path.clone(),
-			reused_existing: true,
-		});
-	}
-
-	let worktree_manager =
-		WorktreeManager::new(project.service_id(), project.repo_root(), project.worktree_root());
-	let worktree = worktree_manager.ensure_worktree_with_hooks(
-		&summary.issue_identifier,
-		false,
-		workflow.frontmatter().execution().workspace_hooks(),
-	)?;
-
-	if worktree.path != summary.worktree_path {
-		eyre::bail!(
-			"planned worktree path `{}` diverged from materialized path `{}` for issue `{}`",
-			summary.worktree_path.display(),
-			worktree.path.display(),
-			summary.issue_identifier
-		);
-	}
-	if worktree.branch_name != summary.branch_name {
-		eyre::bail!(
-			"planned branch `{}` diverged from materialized branch `{}` for issue `{}`",
-			summary.branch_name,
-			worktree.branch_name,
-			summary.issue_identifier
-		);
-	}
-
-	Ok(worktree)
-}
-
-fn spawn_run_once_child(request: SpawnRunOnceChildRequest<'_>) -> Result<Child> {
-	let executable = env::current_exe()?;
-	let lease_preacquired =
-		request.issue_claim_handoff.is_some() || request.dispatch_slot_handoff.is_some();
-	let attempt_request = AttemptRequest {
-		dry_run: false,
-		issue_id: String::from(request.preferred_issue_id),
-		issue_state: String::from(request.preferred_issue_state),
-		initial_issue_state: request.preferred_initial_issue_state.map(String::from),
-		lease_preacquired,
-		#[cfg(unix)]
-		issue_claim_fd: request.issue_claim_handoff.map(AsRawFd::as_raw_fd),
-		#[cfg(not(unix))]
-		issue_claim_fd: None,
-		#[cfg(unix)]
-		dispatch_slot_fd: request.dispatch_slot_handoff.map(AsRawFd::as_raw_fd),
-		#[cfg(not(unix))]
-		dispatch_slot_fd: None,
-		dispatch_slot_index: request.dispatch_slot_index_handoff,
-		dispatch_mode: request.dispatch_mode.into(),
-		run_id: String::from(request.preferred_run_id),
-		attempt_number: request.preferred_attempt_number,
-		retry_budget_base: request.preferred_retry_budget_base,
-		workflow_snapshot: request.workflow.to_markdown()?,
-	};
-	let payload = serde_json::to_vec(&attempt_request)?;
-	let mut command = Command::new(executable);
-
-	command
-		.args(["_attempt", "--config"])
-		.arg(request.config_path)
-		.arg("-")
-		.stdin(Stdio::piped())
-		.stdout(Stdio::inherit())
-		.stderr(Stdio::inherit());
-
-	let mut child = command.spawn()?;
-	let Some(mut stdin) = child.stdin.take() else {
-		let _ = child.kill();
-		let _ = child.wait();
-
-		eyre::bail!("Spawned `_attempt` child without a writable stdin handle.");
-	};
-
-	if let Err(error) = stdin.write_all(&payload) {
-		let _ = child.kill();
-		let _ = child.wait();
-
-		eyre::bail!("Failed to write `_attempt` request payload: {error}");
-	}
-
-	Ok(child)
-}
-
-fn plan_due_retry_run<T>(
-	retry_queue: &mut RetryQueue,
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-) -> Result<RetryDispatchDecision>
-where
-	T: IssueTracker,
-{
-	let now = Instant::now();
-	let Some(first_entry) = retry_queue.next_entry().cloned() else {
-		return Ok(RetryDispatchDecision::Continue);
-	};
-
-	if now < first_entry.ready_at {
-		tracing::debug!(
-			issue_id = first_entry.issue_id,
-			retry_kind = ?first_entry.kind,
-			retry_attempt = first_entry.attempt,
-			"Retry queue is holding the project claim until the next retry is due."
-		);
-
-		return Ok(RetryDispatchDecision::Blocked {
-			excluded_issue_ids: queued_retry_issue_ids(retry_queue),
-		});
-	}
-
-	let mut blocked_issue_id = None;
-
-	for entry in retry_queue.ordered_entries() {
-		if now < entry.ready_at {
-			break;
-		}
-
-		let preferred_issue_state = (entry.kind == RetryKind::Continuation
-			&& !matches!(
-				entry.dispatch_mode,
-				IssueDispatchMode::ReviewRepair | IssueDispatchMode::Closeout
-			))
-		.then_some(workflow.frontmatter().tracker().in_progress_state());
-		let Some(summary) = run_target_issue_once(TargetIssueRunContext {
-			tracker,
-			project,
-			workflow,
-			state_store,
-			issue_id: &entry.issue_id,
-			preferred_issue_state,
-			preferred_initial_issue_state: entry.continuation_initial_issue_state.as_deref(),
-			dry_run: true,
-			lease_preacquired: false,
-			preferred_issue_claim_fd: None,
-			preferred_dispatch_slot_fd: None,
-			preferred_dispatch_slot_index: None,
-			dispatch_mode: entry.dispatch_mode,
-			preferred_run_identity: None,
-			preferred_retry_budget_base: None,
-		})?
-		else {
-			if retry_entry_is_temporarily_blocked(tracker, project, workflow, state_store, &entry)?
-			{
-				blocked_issue_id.get_or_insert_with(|| entry.issue_id.clone());
-
-				continue;
-			}
-
-			clear_retry_schedule_and_release(retry_queue, state_store, &entry.issue_id)?;
-
-			continue;
-		};
-
-		return Ok(RetryDispatchDecision::Dispatch(Box::new(summary)));
-	}
-
-	Ok(blocked_issue_id.map_or(RetryDispatchDecision::Continue, |_issue_id| {
-		RetryDispatchDecision::Blocked { excluded_issue_ids: queued_retry_issue_ids(retry_queue) }
-	}))
-}
-
-fn queued_retry_issue_ids(retry_queue: &RetryQueue) -> Vec<String> {
-	retry_queue.ordered_entries().into_iter().map(|entry| entry.issue_id).collect()
-}
-
-fn stop_daemon_child(child: &mut Child) -> Result<()> {
-	if child.try_wait()?.is_some() {
-		return Ok(());
-	}
-
-	let _ = child.kill();
-	let _ = child.wait();
-
-	Ok(())
-}
